\name{EDH}
\docType{data}
\alias{EDH}
\title{
Epigraphic Database Heidelberg Data Set
}
\description{
This is the Epigraphic Database Heidelberg data set. 
}
\usage{
data("EDH")
}
\format{
A list object of 82464 items (until 20-11-2019) with at least one of the following 42 entries. 

The names in the EDH list are: \code{"ID"}, \code{"commentary"}, \code{"country"}, 
\code{"depth"}, \code{"diplomatic_text"}, \code{"edh_geography_uri"}, \code{"findspot"}, 
\code{"findspot_ancient"}, \code{"findspot_modern"}, \code{"height"}, \code{"id"}, 
\code{"language"}, \code{"last_update"}, \code{"letter_size"}, \code{"literature"}, 
\code{"material"}, \code{"military"}, \code{"modern_region"}, \code{"not_after"}, 
\code{"not_before"}, \code{"person"} (\code{"person_id"}, \code{"nomen"}, \code{"cognomen"}, 
\code{"praenomen"}, \code{"name"}, \code{"gender"}, \code{"status"}, \code{"tribus"}), 
\code{"present_location"}, \code{"province_label"}, \code{"responsible_individual"}, 
\code{"social_economic_legal_history"}, \code{"transcription"}, \code{"trismegistos_uri"}, 
\code{"type_of_inscription"}, \code{"type_of_monument"}, \code{"uri"}, \code{"width"}, 
\code{"work_status"}, and \code{"year_of_find"}. 
}
\source{
https://edh-www.adw.uni-heidelberg.de/data/api
}
%\references{
%}
\keyword{datasets}
