\name{rp}
\docType{data}
\alias{rp}
\title{
Roman provinces
}
\description{
This is a list with Roman provinces as in the Epigraphic Database Heidelberg. 
}
\usage{
data("rp")
}
\format{
A list object of 66 Roman provinces named with \code{"province_label"}.
}
\source{
https://edh-www.adw.uni-heidelberg.de/data/api/terms/province
}
%\references{
%}
\seealso{
\code{\link{get.edh}}, \code{\link{get.edhw}},\code{\link{edhw}}
}
\keyword{datasets}
