\name{EDH}
\docType{data}
\alias{EDH}
\title{
Epigraphic Database Heidelberg Data Set
}
\description{
This is the Epigraphic Database Heidelberg data set. 
}
\usage{
data("EDH")
}
\format{
A list object of 82464 items with 28 entries each. 

The names in the EDH list are \code{"ID"}, \code{"commentary"}, \code{"country"}, 
\code{"depth"}, \code{"diplomatic_text"}, \code{"edh_geography_uri"}, 
\code{"findspot_ancient"}, \code{"findspot_modern"}, \code{"height"}, \code{"id"}, 
\code{"language"}, \code{"last_update"}, \code{"letter_size"}, \code{"literature"}, 
\code{"material"}, \code{"modern_region"}, \code{"not_after"}, \code{"not_before"}, 
\code{"people"}, \code{"province_label"}, \code{"responsible_individual"}, 
\code{"transcription"}, \code{"trismegistos_uri"}, \code{"type_of_inscription"}, 
\code{"type_of_monument"}, \code{"uri"}, \code{"width"}, and \code{"work_status"}. 

}
\source{
https://edh-www.adw.uni-heidelberg.de/data/api
}
%\references{
%}
\keyword{datasets}
