\name{rp}
\docType{data}
\alias{rp}
\title{
Roman province names and acronyms as in EDH dataset
}
\description{
This is a list with Roman province names and acronyms as in the Epigraphic Database Heidelberg recorded in \code{EDH} dataset. 
}
\usage{
data("rp")
}
\format{
A list object of 66 Roman provinces names and acronyms as in \code{"province_label"} in \code{EDH} dataset.
}
\source{
https://edh-www.adw.uni-heidelberg.de/data/api/terms/province
}
%\references{
%}
\seealso{
\code{\link{get.edh}}, \code{EDH}, \code{\link{edhw}}, \code{retn}, \code{rpmp}, 
}
\keyword{datasets}
